requirements.txt
Django>=3.1.4,<3.2.0
djangorestframework>=3.12.0,<3.13.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
django-extensions>=3.1.1,<3.2.0

mysqlclient>=2.0.3,<2.1.0
psycopg2>=2.8.6,<2.9.0
djongo>=1.3.4,<1.4.0

flake8>=3.6.0,<3.7.0